Rename mock file names and update imports
